Size_t count for restore.

+++ src/skia/core/canvas.d
@@ -252,10 +252,10 @@ public:
   /****************************************
    * Stub
    */
-  int save(SaveFlags flags = SaveFlags.MatrixClip) {
+  size_t save(SaveFlags flags = SaveFlags.MatrixClip) {
     return this.internalSave(flags);
   }
-  private final int internalSave(SaveFlags flags) {
+  private final size_t internalSave(SaveFlags flags) {
     this.mcRecs ~= this.curMCRec;
     return this.mcRecs.length - 1;
   }
@@ -272,7 +272,7 @@ public:
     assert(this.mcRecs.length > 0);
     this.mcRecs = this.mcRecs[0 .. $-1];
   }
-  void restoreCount(uint sc) {
+  void restoreCount(size_t sc) {
     assert(sc <= this.mcRecs.length);
     this.mcRecs = this.mcRecs[0 .. sc];
   }
